+++ src/fields/OtpField/index.ts
@@ -1,1 +1,2 @@
 export * from './OtpField'
+export { otpFieldClasses } from './OtpFieldControl'