+++ src/theme/augmentation.ts
@@ -6,6 +6,8 @@ import type { WizardStepperProps } from '../Wizard/WizardStepper'
 import type { WizardNavProps } from '../Wizard/WizardNav'
 import type { ReadOnlyFieldProps } from '../fields/ReadOnlyField'
 import type { NumberFieldControlProps } from '../fields/NumberField/NumberFieldControl'
+import type { OtpFieldControlProps } from '../fields/OtpField/OtpFieldControl'
+import type { FileFieldProps } from '../fields/FileField'
 
 declare module '@mui/material/styles' {
   interface ComponentsPropsList {
@@ -16,6 +18,8 @@ declare module '@mui/material/styles' {
     EzWizardNav: Partial<WizardNavProps>
     EzReadOnlyField: Partial<ReadOnlyFieldProps>
     EzNumberField: Partial<NumberFieldControlProps>
+    EzOtpField: Partial<OtpFieldControlProps>
+    EzFileField: Partial<FileFieldProps>
   }
 
   interface ComponentNameToClassKey {
@@ -26,6 +30,8 @@ declare module '@mui/material/styles' {
     EzWizardNav: 'root' | 'prev' | 'next' | 'submit'
     EzReadOnlyField: 'root' | 'header' | 'label' | 'value' | 'edit'
     EzNumberField: 'root' | 'steppers' | 'increment' | 'decrement'
+    EzOtpField: 'root' | 'helperText'
+    EzFileField: 'root' | 'fileList'
   }
 
   interface Components<Theme = unknown> {
@@ -56,6 +62,14 @@ declare module '@mui/material/styles' {
     EzNumberField?: {
       defaultProps?: ComponentsProps['EzNumberField']
       styleOverrides?: ComponentsOverrides<Theme>['EzNumberField']
+    }
+    EzOtpField?: {
+      defaultProps?: ComponentsProps['EzOtpField']
+      styleOverrides?: ComponentsOverrides<Theme>['EzOtpField']
+    }
+    EzFileField?: {
+      defaultProps?: ComponentsProps['EzFileField']
+      styleOverrides?: ComponentsOverrides<Theme>['EzFileField']
     }
   }
 }